Initial static prototype of the landing page for KISSmetrics.
I'm pleasantly nervous about putting this much text on a landing page.
It's a bit counterintuitive and that's why I have a weird hunch that it
could do really well. If it tanks, it still might not be a bad page, we
might just need more of this kind of educational content so that people
can click around a few times and get ideas for the kinds of things they
can do.

At the very least, this is a good warmup for redesigning the blog.

Design:
* it's a single column design like Om Malik's, 37signals,
UsabilityPost, or like if you used Instapaper or Readability plugins to
strip a blog to just the content.
* I modified the embedded Zaps lightly, to make the more square and
flat.

Responsive Layout:
* both the browser width and height are triggers for reformatting the
layout, so that the white text content is above the fold on desktops
and tablets.
* the header and content are formatted for phones and tablet.
* most of the body content is too, with the exception of the embedded
Zaps on phones. They're *almost* usable and it'd be rad to get those
working too, but I'm not sure how mobile-friendly the rest of the app
is, how much of a project it would be to tackle it, or how many mobile
visitors we get.

Typography:
* switched to Zapier's Open Sans from the Helvetica Neue in my mocks.
* used a golden ratio for font sizes, which is a Fibonacci-based scale
for balancing the different types of writing. E.g. on desktop, it
starts with 84px for headlines, 52px for subheaders, 32px for large
text, and 20px for body text. Each one is 1.61 times larger than the
next.
* added two CSS classes to increase legibility, -webkit-font-smoothing
and text-rendering. Open Sans is a bit of a heavy text and this
lightens up the body copy so it doesn't look so jaggy.

Buttons:
* sounds strange but I'm not a fan of UI consistency in some areas,
especially buttons. I can explain more in-depth at some point.
* varying button styles allows for better emphasis of primary and
secondary actions from a UX perspective.
* it also helps with visual design because elements can clash.
* examples of varying styles on a single page - apple/mavericks.
flowdock?

Language:
* for first time readers, it should help to explain what some of our
internal jargon is, e.g. the little caption on "what is a zap?" below
the first one.
* I'm a grammar nerd about "setup" being a noun vs "set up" being an
action :)
* same with login/log in, signup/sign up.

Links:
* I removed the links to other tools at the bottom of the post, which
is great for a blog post but want to lean towards a really sticky
funnel for landing pages.